\name{RohMut-package}
\alias{RohMut-package}
\alias{RohMut}
\docType{package}
\title{
  The package allows you to estimate the sex-specific mutation rate based on runs of homozygosity segments from close common ancestor.
}
\description{
  The package uses mixture of poissons in order to estimate the sex-specific mutation rate, for ROH segments coming from offspring of either first or second cousins.
}
\details{
  See estimate_EM for more details.
}
\author{
Liraz Klausner
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
  \dontrun{
  }
}
